Added TMS/SWDIO direction pin & made SRST push pull.
This version uses a direction control level shifters. We need to control
the direction of the TMS/SWDIO pin. Additionally because we translate
SRST using the level shifter we can not implement open drain control of
that pin. For this hardware version PushPull or disconnecting the pin
seem like the only option. :(