Level shifter bypass modifications.
Using PB14 instead of PA2 for SRST. Requires a hardware modification on
the BMPM V2.1a that disconnects the level shifter from the xRST net and
connects the PB14 pin to the xRST. The level shifter is only push pull
so the previous code change to Push/Pull operation was needed. This can
cause problems with chips that pull the RST pin by themselves when
resetting. This is why we decided to connect the pin to a 5V tolerant
GPIO on the stm32, and bypass the levelshifter.